investment_dashboard/requirements_investment.txt
# 投資ダッシュボード 依存ライブラリ
# pip install -r requirements_investment.txt

# 株価・金融データ
yfinance>=0.2.36

# RSS・スクレイピング
feedparser==6.0.11
beautifulsoup4>=4.12.0
requests>=2.31.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# テクニカル指標
pandas-ta>=0.3.14b

# Web UI
Flask>=3.0.0

# 設定ファイル
pyyaml>=6.0.1
